Improve resource comsumption
