+++ CMakeLists.txt
@@ -36,12 +36,14 @@ add_custom_target(steep
 )
 
 add_custom_target(ruby
-  COMMAND ${RBC} -Bkeyboard keyboard.rb
-  COMMAND ${RBC} -Btud tud.rb
-  COMMAND ${RBC} -Brgb rgb.rb
+  COMMAND ${RBC} -Bcore           core.rb
+  COMMAND ${RBC} -Bkeyboard       keyboard.rb
+  COMMAND ${RBC} -Brgb            rgb.rb
   COMMAND ${RBC} -Brotary_encoder rotary_encoder.rb
-  COMMAND ${RBC} -Bsandbox sandbox.rb
-  COMMAND ${RBC} -Bkeymap ${PROJECT_BINARY_DIR}/../keymap.rb
+  COMMAND ${RBC} -Btud            tud.rb
+  COMMAND ${RBC} -Brgb_task       rgb_task.rb
+  COMMAND ${RBC} -Bsandbox        sandbox.rb
+  COMMAND ${RBC} -Bkeymap         ${PROJECT_BINARY_DIR}/../keymap.rb
   COMMAND mv ${PROJECT_BINARY_DIR}/../keymap.c ./
   WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/src/ruby/lib
 )
