+++ src/main.c
@@ -18,11 +18,15 @@
 #include "ruby_mode.h"
 
 /* ruby */
+/* models */
+#include "ruby/lib/core.c"
 #include "ruby/lib/keyboard.c"
-#include "ruby/lib/keymap.c"
-#include "ruby/lib/tud.c"
-#include "ruby/lib/rgb.c"
 #include "ruby/lib/rotary_encoder.c"
+#include "ruby/lib/rgb.c"
+/* tasks */
+#include "ruby/lib/tud.c"
+#include "ruby/lib/rgb_task.c"
+#include "ruby/lib/keymap.c"
 #include "ruby/lib/sandbox.c"
 
 void
@@ -37,9 +41,26 @@ c_rand(mrb_vm *vm, mrb_value *v, int argc)
   SET_INT_RETURN(rand());
 }
 
-#define MEMORY_SIZE (1024*200)
+#define MEMORY_SIZE (1024*220)
 
 static uint8_t memory_pool[MEMORY_SIZE];
+
+void
+mrbc_load_model(const uint8_t *mrb)
+{
+  mrbc_vm *vm = mrbc_vm_open(NULL);
+  if( vm == 0 ) {
+    console_printf("Error: Can't open VM.\n");
+    return;
+  }
+  if( mrbc_load_mrb(vm, mrb) != 0 ) {
+    console_printf("Error: Illegal bytecode.\n");
+    return;
+  }
+  mrbc_vm_begin(vm);
+  mrbc_vm_run(vm);
+  mrbc_raw_free(vm);
+}
 
 mrbc_tcb *tcb_rgb; /* from ws2812.h */
 
@@ -56,10 +77,13 @@ int main() {
   WS2812_INIT();
   ROTARY_ENCODER_INIT();
   RUBY_MODE_INIT();
-//  tcb_rgb = mrbc_create_task(rgb, 0);
-  mrbc_create_task(rotary_encoder, 0);
+  mrbc_load_model(core);
+  mrbc_load_model(rgb);
+  mrbc_load_model(rotary_encoder);
+  mrbc_load_model(keyboard);
+  mrbc_create_task(sandbox, 0);
   mrbc_create_task(tud, 0);
-  mrbc_create_task(keyboard, 0);
+  tcb_rgb = mrbc_create_task(rgb_task, 0);
   mrbc_create_task(keymap, 0);
   mrbc_run();
   return 0;
