+++ src/ruby/lib/keyboard.rb
@@ -1,17 +1,3 @@
-class Float
-  def modulo(right)
-    left = self.to_f
-    while left > right
-      left -= right
-    end
-    left
-  end
-  def ceil_to_i
-    n = self.to_i
-    (self > n) ? (n + 1) : n
-  end
-end
-
 class Keyboard
   GPIO_OUT = 1
   GPIO_IN  = 0
