+++ src/ruby/lib/rgb.rb
@@ -1,27 +1,3 @@
-def hsv2rgb(h, s, v)
-  s /= 100.0
-  v /= 100.0
-  c = v * s
-  x = c * (1 - ((h / 60.0).modulo(2) - 1).abs)
-  m = v - c
-  rgb = if h < 60
-          [c, x, 0]
-        elsif h < 120
-          [x, c, 0]
-        elsif h < 180
-          [0, c, x]
-        elsif h < 240
-          [0, x, c]
-        elsif h < 300
-          [x, 0, c]
-        else
-          [c, 0, x]
-        end
-  ((rgb[0] + m) * 255).ceil_to_i << 16 |
-    ((rgb[1] + m) * 255).ceil_to_i << 8 |
-    ((rgb[2] + m) * 255).ceil_to_i
-end
-
 class RGB
   def initialize(pin, underglow_size, backlight_size, is_rgbw)
     @fifo = Array.new
@@ -80,36 +56,3 @@ class RGB
   end
 
 end
-
-# Suspend itself until being resumed in Keyboard#start_rgb
-suspend_task
-
-hue = 0
-
-case $rgb.effect
-when :rainbow
-  step = 360.0 / $rgb.pixel_size
-  $rgb.pixel_size.times do |i|
-    $rgb.set_pixel_at(i, hsv2rgb(i * step, 100, 12.5))
-  end
-when :breathing
-  hue = 0
-end
-
-while true
-  if $rgb.key?
-    case $rgb.action
-    when :thunder
-      $rgb.thunder
-    end
-  end
-  case $rgb.effect
-  when :rainbow
-    $rgb.rotate
-  when :breathing
-    $rgb.fill(hsv2rgb(hue, 100, 12.5))
-    hue >= 360 ? hue = 0 : hue += 10
-  end
-  $rgb.show
-end
-
